Instrucoes decisao

+++ Ficha1/Ex2a.c
@@ -10,19 +10,19 @@ void main()
     float subtracao;
     float multi;
     float divi;
-    float resto;
+    int resto;
 
     printf("Primeiro numero inteiro:");
-    scanf("%f",&num1);
+    scanf("%d",&num1);
 
     printf("Segundo numero inteiro:");
-    scanf("%f",&num2);
+    scanf("%d",&num2);
 
     adicao = num1 + num2;
     subtracao = num1 - num2;
     multi = num1 * num2;
     divi = num1 / num2;
     resto = num1 % num2;
-    printf("%f",resto);
+    printf("%d",resto);
 
 }
